requirements.txt
homoglyphs==2.0.4
Pillow==9.3.0
python-socketio==5.9.0